refactor: restore typing for non-backports
