+++ cibuildwheel/options.py
@@ -22,12 +22,13 @@ else:
 
 from packaging.specifiers import SpecifierSet
 
-from ._compat.typing import PLATFORMS, Literal, NotRequired, PlatformName, TypedDict
+from ._compat.typing import Literal, NotRequired, TypedDict
 from .architecture import Architecture
 from .environment import EnvironmentParseError, ParsedEnvironment, parse_environment
 from .logger import log
 from .oci_container import ContainerEngine
 from .projectfiles import get_requires_python_str
+from .typing import PLATFORMS, PlatformName
 from .util import (
     MANYLINUX_ARCHS,
     MUSLLINUX_ARCHS,
